requirements.txt
Django==1.8.5
argparse==1.2.1
dj-database-url==0.3.0
django-cors-headers==1.1.0
djangorestframework==3.2.4
psycopg2==2.6.1
requests==2.8.0
wsgiref==0.1.2
